+++ iris/iris_facade_test.go
@@ -1,6 +1,8 @@
 package iris_test
 
 import (
+	"time"
+
 	"github.com/park285/iris-client-go/iris"
 )
 
@@ -11,4 +13,6 @@ var (
 	_ iris.BotClient    = (*iris.RebindingClient)(nil)
 	_ iris.Client       = (*iris.RebindingClient)(nil)
 	_ iris.KaringClient = (*iris.RebindingClient)(nil)
+
+	_ = iris.RebindingClientConfig{ResolveInterval: time.Second}
 )